requirements-ml.txt
# ML-specific requirements for FinPulse
pandas>=1.5.0
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0
matplotlib>=3.7.0
packaging>=23.0
pyyaml>=6.0
openpyxl>=3.1.0